need to figure out how to refactor views.py at a later date
