+++ backend/ht_buses_app/refactor/users/user_delete.py
@@ -10,7 +10,7 @@ def user_delete(request):
     data = {}
     id = request.query_params["id"]
     try:
-        user_object =  User.objects.filter(first_name = reqBody['first_name'], last_name =reqBody['last_name'], email = reqBody['email'])
+        user_object =  User.objects.get(pk=id)
         user_object.delete()
         data["message"] = "user successfully deleted"
         result = {"data" : data}
