+++ backend/ht_buses_app/tests/routes/test_routes.py
@@ -1,7 +1,7 @@
 import json
 from rest_framework.test import APIClient, APITestCase
 from django.core.exceptions import ValidationError
-from ..models import UserManager, School, Route
+from ...models import UserManager, School, Route
 
 
 class Test_routes(APITestCase):
